refactor(handlers): extract 8 provider families from imageGeneration.ts (#3594)
Break the 3776-line imageGeneration handler into provider-specific modules:
- imageGeneration/providers/chatgptWeb.ts
- imageGeneration/providers/comfyUI.ts
- imageGeneration/providers/haiper.ts
- imageGeneration/providers/hyperbolic.ts
- imageGeneration/providers/ideogram.ts
- imageGeneration/providers/imagen3.ts
- imageGeneration/providers/leonardo.ts
- imageGeneration/providers/sdWebUI.ts

Pure extraction, no behavioral changes. Supersedes #4609. Part of #3594.
